refactor: rewrite code according to dry
