+++ async-race/src/app/components/winners-leader-board/_winners-leader-board.scss
@@ -1,6 +1,8 @@
 .winners-leader-board {
   height: 600px;
+  margin: 30px;
+
   font-size: 3rem;
   color: $font-color;
-  text-align: center 
+  text-align: center;
 }
